+++ business/ccard/account/GoldCard.java
@@ -3,6 +3,9 @@ package edu.mum.cs.cs525.labs.exercises.project.business.ccard.account;
 import edu.mum.cs.cs525.labs.exercises.project.business.framework.Account;
 import edu.mum.cs.cs525.labs.exercises.project.business.framework.InterestStrategy;
 
+import static edu.mum.cs.cs525.labs.exercises.project.business.ccard.ApplicationConfiguration.BRONZE_CARD;
+import static edu.mum.cs.cs525.labs.exercises.project.business.ccard.ApplicationConfiguration.GOLD_CARD;
+
 public class GoldCard extends Account {
 
     String name,expDate,type;
@@ -43,6 +46,17 @@ public class GoldCard extends Account {
     @Override
     public String getOwnershipType() {
         return ownershipType;
+    }
+
+    @Override
+    protected double calculateNewBalance() {
+        return getPreviousBalance() - getTotalCredits() + getTotalCharges() + GOLD_CARD.getMonthlyInterestRate() * (getPreviousBalance() - getTotalCredits());
+
+    }
+
+    @Override
+    protected double calculateTotalDue(double newBalance) {
+        return GOLD_CARD.getMinimumPayment() * newBalance;
     }
 
     public void setOwnershipType(String ownershipType) {
