+++ ui/ccard/JDialogGenBill.java
@@ -3,26 +3,13 @@ package edu.mum.cs.cs525.labs.exercises.project.ui.ccard;
 		A basic implementation of the JDialog class.
 */
 
-import edu.mum.cs.cs525.labs.exercises.project.business.bank.BankAccountType;
-import edu.mum.cs.cs525.labs.exercises.project.business.bank.BankApplicationImpl;
-import edu.mum.cs.cs525.labs.exercises.project.business.bank.account.CheckingAccount;
-import edu.mum.cs.cs525.labs.exercises.project.business.bank.template.BankAccountReport;
 import edu.mum.cs.cs525.labs.exercises.project.business.ccard.CreditCardAccountType;
-import edu.mum.cs.cs525.labs.exercises.project.business.ccard.CreditCardApplicationImpl;
-import edu.mum.cs.cs525.labs.exercises.project.business.ccard.account.BronzeCard;
-import edu.mum.cs.cs525.labs.exercises.project.business.ccard.account.GoldCard;
-import edu.mum.cs.cs525.labs.exercises.project.business.ccard.account.SilverCard;
-import edu.mum.cs.cs525.labs.exercises.project.business.ccard.template.BronzeCardReport;
-import edu.mum.cs.cs525.labs.exercises.project.business.ccard.template.GoldCardReport;
-import edu.mum.cs.cs525.labs.exercises.project.business.ccard.template.SilverCardReport;
 import edu.mum.cs.cs525.labs.exercises.project.business.framework.Account;
 import edu.mum.cs.cs525.labs.exercises.project.business.framework.ApplicationFacade;
-import edu.mum.cs.cs525.labs.exercises.project.business.framework.Report;
 
-import java.awt.*;
-import java.util.List;
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
+import java.awt.*;
 
 public class JDialogGenBill extends javax.swing.JDialog {
     private ApplicationFacade<CreditCardAccountType> applicationFacade;
@@ -54,20 +41,8 @@ public class JDialogGenBill extends javax.swing.JDialog {
             String accountNumber = (String) model.getValueAt(selection, 1);
             Account acc = applicationFacade.getAccount(accountNumber);
             if (acc != null) {
-                Report reportClass;
-                if (acc instanceof GoldCard) {
-                    reportClass = new GoldCardReport(acc);
-                } else if (acc instanceof SilverCard) {
-                    reportClass = new SilverCardReport(acc);
-                } else if (acc instanceof BronzeCard) {
-                    reportClass = new BronzeCardReport(acc);
-                } else if (acc instanceof CheckingAccount) {
-                    reportClass = new BankAccountReport(acc);
-                } else {
-                    reportClass = new BankAccountReport(acc);
-                }
                 System.out.println("Details are being generated");
-                String report = reportClass.generate();
+                String report = acc.generate();
                 JTextField1.setText(report);
             } else {
                 JTextField1.setText("Account not found.");
